requirements.txt
python-docx
openai
langchain
faiss-cpu
tqdm
python-telegram-bot==20.3
tiktoken
matplotlib
pandas
